incr.comp.: Remove some code duplication around TyDecoder by factoring things into a macro.
